chore: refactor autocomplete and address comments
